constants refactoring
