+++ backend-services/detection/core/detection.py
@@ -1,7 +1,6 @@
 import ipaddress
 import json as classic_json
 import multiprocessing as mp
-import os
 import re
 import time
 from datetime import datetime
@@ -18,13 +17,18 @@ from artemis_utils import exception_handler
 from artemis_utils import get_hash
 from artemis_utils import get_ip_version
 from artemis_utils import get_logger
-from artemis_utils import RABBITMQ_URI
-from artemis_utils import REDIS_HOST
-from artemis_utils import REDIS_PORT
-from artemis_utils import RPKI_VALIDATOR_ENABLED
-from artemis_utils import RPKI_VALIDATOR_HOST
-from artemis_utils import RPKI_VALIDATOR_PORT
-from artemis_utils import TEST_ENV
+from artemis_utils.constants import DATABASE_HOST
+from artemis_utils.constants import HEALTH_CHECK_TIMEOUT
+from artemis_utils.constants import NOTIFIER_HOST
+from artemis_utils.constants import PREFIXTREE_HOST
+from artemis_utils.envvars import RABBITMQ_URI
+from artemis_utils.envvars import REDIS_HOST
+from artemis_utils.envvars import REDIS_PORT
+from artemis_utils.envvars import REST_PORT
+from artemis_utils.envvars import RPKI_VALIDATOR_ENABLED
+from artemis_utils.envvars import RPKI_VALIDATOR_HOST
+from artemis_utils.envvars import RPKI_VALIDATOR_PORT
+from artemis_utils.envvars import TEST_ENV
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
 from artemis_utils.redis import ping_redis
@@ -50,10 +54,6 @@ shared_memory_locks = {"data_worker": mp.Lock()}
 
 # global vars
 SERVICE_NAME = "detection"
-NOTIFIER_HOST = "notifier"
-PREFIXTREE_HOST = "prefixtree"
-DATABASE_HOST = "database"
-REST_PORT = int(os.getenv("REST_PORT", 3000))
 HIJACK_DIM_COMBINATIONS = [
     ["S", "0", "-", "-"],
     ["S", "0", "-", "L"],
@@ -72,8 +72,6 @@ HIJACK_DIM_COMBINATIONS = [
     ["Q", "0", "-", "L"],
 ]
 DATA_WORKER_DEPENDENCIES = [PREFIXTREE_HOST, DATABASE_HOST, NOTIFIER_HOST]
-# need to move to utils
-HEALTH_CHECK_TIMEOUT = 5
 
 
 # need to move this to utils
