+++ backend-services/notifier/core/notifier.py
@@ -8,7 +8,9 @@ from typing import NoReturn
 import requests
 import ujson as json
 from artemis_utils import get_logger
-from artemis_utils import RABBITMQ_URI
+from artemis_utils.constants import CONFIGURATION_HOST
+from artemis_utils.envvars import RABBITMQ_URI
+from artemis_utils.envvars import REST_PORT
 from artemis_utils.logging import hijack_log_field_formatter
 from artemis_utils.rabbitmq import create_exchange
 from artemis_utils.rabbitmq import create_queue
@@ -53,8 +55,6 @@ shared_memory_locks = {"data_worker": mp.Lock(), "config_timestamp": mp.Lock()}
 
 # global vars
 SERVICE_NAME = "notifier"
-CONFIGURATION_HOST = "configuration"
-REST_PORT = int(os.getenv("REST_PORT", 3000))
 
 
 def configure_notifier(msg, shared_memory_manager_dict):
